Preserve declared numeric width across introspect, invent and bind
Three carriers were being re-derived from an ambiguous keyword instead of the
declared source domain:

- Oracle ANSI FLOAT(p) is NUMBER-backed binary precision (bare = 126 binary
  digits) and was read, and rematerialized, as IEEE BINARY_DOUBLE — an
  Oracle-to-Oracle column silently dropped to a 53-bit mantissa
- SQL Server int introspected as BIGINT, so a 32-bit column was recreated
  64-bit; it now names the width the way PostgreSQL int4 does
- the SQLAlchemy bind re-widened an INT wire to BigInteger after
  materialization had already resolved the ambiguity against the named engine

Bare stamps keep the never-narrower default: an unnamed INTEGER may be an
Oracle NUMBER(38) that overflows an int32 column at row 1.

Tests that asserted the older spellings are updated with the reasoning:
MySQL FLOAT is IEEE-32 (FLOAT32), Snowflake FLOAT is IEEE-64.

+++ apps/api/tests/test_schema_introspect_specialty.py
@@ -43,12 +43,16 @@ def test_pg_mysql_bq_sf_preserve_ieee_float():
     assert _pg_to_logical("float8") == "DOUBLE PRECISION"
     assert _pg_to_logical("numeric") == "DECIMAL"
     assert _mysql_to_logical("double") == "DOUBLE"
-    assert _mysql_to_logical("float") == "FLOAT"
+    # MySQL FLOAT is IEEE-32; the bare FLOAT keyword means IEEE-64 on
+    # PostgreSQL/SQL Server/Snowflake, so the width is named explicitly.
+    assert _mysql_to_logical("float") == "FLOAT32"
     assert _mysql_to_logical("decimal(10,2)") == "DECIMAL(10,2)"
     assert _bq_to_logical("FLOAT64") == "FLOAT64"
     # BIGNUMERIC stays distinct from NUMERIC/DECIMAL (76,38 vs 38,9 contract).
     assert _bq_to_logical("BIGNUMERIC") == "BIGNUMERIC"
-    assert _sf_to_logical("FLOAT") == "FLOAT"
+    # Snowflake FLOAT / FLOAT8 / DOUBLE are one IEEE-64 type — naming the width
+    # keeps a 32-bit destination from being invented from the spelling.
+    assert _sf_to_logical("FLOAT") == "DOUBLE"
     assert _sf_to_logical("FLOAT8") == "DOUBLE"
     assert _sf_to_logical("NUMBER(38,10)") == "DECIMAL(38,10)"
     assert _sample_logical_type(1.5) == "DOUBLE"
